% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-characteristic.R
\name{geograbbi.extract.characteristics}
\alias{geograbbi.extract.characteristics}
\title{Extract a characteristic field in GEO sample data}
\usage{
geograbbi.extract.characteristics(sample.frame,
  chr.field.str = "characteristics")
}
\arguments{
\item{sample.frame}{a data.frame of descriptives for a GEO experiment (i.e GSE) 
say as returned by \code{geograbbi.get.samples}}

\item{chr.field.str}{string that identifies the names of the characteristic
columns to be extracted}
}
\description{
Pattern searches (greps) for a particular pattern in a data object with 
sample
}
