% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{geograbbi.get.data}
\alias{geograbbi.get.data}
\title{Get GEO data}
\usage{
geograbbi.get.data(filename, path = NULL, FUN = geograbbi.read.gse.matrix)
}
\arguments{
\item{filename}{}
}
\description{
Grabs the expression/microarray/etc. data contained in the GEO series
 matrix files
}
