% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{geograbbi.read.table}
\alias{geograbbi.read.table}
\title{Read a very large table}
\usage{
geograbbi.read.table(file, header = F, chunk.size = 21234, verbose = F,
  debug = F, ...)
}
\arguments{
\item{file}{A character connection or filename.}

\item{header}{A logical indicating if the file contains a header
(default: \code{FALSE}).}

\item{chunk.size}{Size of partitions to read in bytes
(default: 21234).}

\item{verbose}{A logical indicating whether or not to print
status updates (default: \code{FALSE}).}

\item{debug}{A logical indicating whether or not to run
\code{\link{browser}()} after attempting to read the file partitions.}

\item{...}{Additional arguments passed to \code{\link{read.table}()}.}
}
\value{
A data frame as if \code{\link{read.table}()} had been used.
}
\description{
\code{\link{read.table}()} is extremely memory inefficient and slow.
This function somewhat resolves this by applying \code{\link{read.table}()}
multiple times to partitions of the table.
}
