% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage-db.R
\name{geograbbi.replace.GEOmetadb.sqlite}
\alias{geograbbi.replace.GEOmetadb.sqlite}
\title{Update the GEOmetadb sql directory}
\usage{
geograbbi.replace.GEOmetadb.sqlite(archive.path = ".")
}
\arguments{
\item{archive.path}{path where GEOmetadb.sqlite will be archived}
}
\description{
Moves the current GEOmetadb.sqlite database to an archived directory of 
your choice and then downloads a new version of this file
}
