% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-series-files.R
\name{geograbi.download.supplementary.file}
\alias{geograbi.download.supplementary.file}
\title{Download a GEO supplementary file}
\usage{
geograbi.download.supplementary.file(path = ".", gse = NULL,
  filename = NULL, url = NULL)
}
\arguments{
\item{path}{path for saving downloaded GEO series matrix files}

\item{gse}{character of gse of interest}

\item{filename}{name of specific supplemental file from gse to downloads}

\item{url}{optional manual url supply for non-ftp based file downloads}
}
\description{
Takes a GSE and a corresponding GEO supplementary file name
and downloads that file to a specified directory
}
