% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{geograbi.get.data}
\alias{geograbi.get.data}
\title{Get GEO data}
\usage{
geograbi.get.data(filename, path = NULL,
  FUN = geograbi.read.gse.matrix)
}
\arguments{
\item{filename}{}
}
\description{
Grabs the expression/microarray/etc. data contained in the GEO series
 matrix files
}
