% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{geograbi.list.supplementary.files}
\alias{geograbi.list.supplementary.files}
\title{List the GEO supplementary files available for a GSE experiment}
\usage{
geograbi.list.supplementary.files(filename, FUN = geograbi.get.series)
}
\description{
List the GEO supplementary files available for a GSE experiment
}
