% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{geograbi.read.csv}
\alias{geograbi.read.csv}
\title{Read a very large csv file}
\usage{
geograbi.read.csv(file, header = TRUE, sep = ",", quote = "\\"",
  dec = ".", fill = TRUE, comment.char = "", verbose = F, debug = F,
  chunk.size = 21234, ...)
}
\description{
This function is a wrapper for \code{\link{geograbi.read.table}()}
for reading csv files.
}
\examples{
\donttest{
## download a DNA methylation dataset
dataset <- geograbi.read.gse.matrix("GSE116339")
## dataset$data is empty because sometimes GEO
## stores the genomic data in a supplementary file
## instead of in the series matrix file

## download the supplementary file
filename <- geograbi.download.supplementary.file(gse="GSE116339",
                    filename="GSE116339_Processed_Matrix.csv.gz")
## 15 minutes

## load the file
dataset$data <- geograbi.read.csv(filename)
## 35 minutes

## make the first column the rownames (features)
## and convert to a matrix
sites <- as.character(dataset$data[,1])
dataset$data <- as.matrix(dataset$data[,-1])
rownames(dataset$data) <- sites

## reorder the samples according to data matrix columns
colnames(dataset$data) <- sub("X", "", colnames(dataset$data))
idx <- match(colnames(dataset$data), dataset$samples$title)
dataset$samples <- dataset$samples[idx,]
}

}
