% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{geograbi.read.gse.matrix}
\alias{geograbi.read.gse.matrix}
\title{Load a GSE matrix file from GEO}
\usage{
geograbi.read.gse.matrix(filename, data = TRUE)
}
\arguments{
\item{filename}{A character string that is a GSE identifier,
a url for GSE matrix file online, or a filename for a local
GSE matrix file.}

\item{data}{If \code{TRUE}, then load the data matrix,
otherwise load only the sample information.}
}
\value{
A list with three elements:
\itemize{
\item{"samples"}{The sample data frame with one row per sample.}
\item{"series"}{The series vector with information on the entire gse 
series record}
\item{"data"}{The data matrix with one row per feature and one per sample.}
}
}
\description{
Load a GSE matrix file from GEO
}
