% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{geograbi.read.table}
\alias{geograbi.read.table}
\title{Read a very large table}
\usage{
geograbi.read.table(file, header = F, chunk.size = NA, verbose = F, ...)
}
\arguments{
\item{file}{A character connection or filename.}

\item{header}{A logical indicating if the file contains a header
(default: \code{FALSE}).}

\item{chunk.size}{Deprecated parameter.
(default: NA).}

\item{verbose}{A logical indicating whether or not to print
status updates (default: \code{FALSE}).}

\item{...}{Additional arguments passed to \code{\link{data.table::fread}()}.}
}
\value{
A data frame as if \code{\link{read.table}()} had been used.
}
\description{
This function used to be an attempt to adapt \code{\link{read.table}()}
for large files. It is now a wrapper for \code{\link{data.table::fread}()}.
}
