% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gds.R
\name{geograbi.retrieve.datasets}
\alias{geograbi.retrieve.datasets}
\title{Retrieve GEO dataset summaries}
\usage{
geograbi.retrieve.datasets(gpl = NULL, retmax = 10000)
}
\arguments{
\item{gpl}{Restrict summaries to those datasets that use
the platforms listed.  Default is NULL, no platform restriction.}

\item{retmax}{Maximum number of dataset summaries to retrieve
in a single step.}
}
\value{
Data frame with one row per dataset.
}
\description{
Retrieve GEO dataset summaries
}
\examples{
  ## all datasets in GEO
  datasets <- geograbi.retrieve.datasets()  ## about 4min

  ## all Illumina Infinium 27K DNA methylation datasets
  datasets <- geograbi.retrieve.datasets(c("GPL8490")) ## 2sec

  ## all Illumina Infinium 27K/450K/EPIC DNA methylation datasets
  datasets <- geograbi.retrieve.datasets(c("GPL8490", "GPL13534", "GPL21145")) ## 5sec

}
