% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage-db.R
\name{geograbr.connect.db}
\alias{geograbr.connect.db}
\title{Open connection with GEOmetadb}
\usage{
geograbr.connect.db(db.path = ".")
}
\arguments{
\item{db.path}{path where GEOmetadb.sqlite is stored}
}
\description{
A wrapper for connecting to the GEOmetadb sql database to allow 
dbListFields into sql queries
}
