% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-series-files.R
\name{geograbr.download.series.files}
\alias{geograbr.download.series.files}
\title{Download GEO series matrix files}
\usage{
geograbr.download.series.files(path = ".", gses, gpls = NULL)
}
\arguments{
\item{path}{path for saving downloaded GEO series matrix files}

\item{gses}{vector of gses to download from GEO}

\item{gpls}{vector of platform gpl codes with same length as \code{gses}}
}
\description{
Takes a vector of GSEs, say as returned by \code{geo.query.db.gses}, and 
their corresponding gpl codes
and downloads the GEO series matrix files to a specified directory
}
