% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download-series-files.R
\name{geograbr.download.supplementary.file}
\alias{geograbr.download.supplementary.file}
\title{Download a GEO supplementary file}
\usage{
geograbr.download.supplementary.file(path = ".", gse, sup.file.name)
}
\arguments{
\item{path}{path for saving downloaded GEO series matrix files}

\item{gse}{character of gse of interest}

\item{sup.file.name}{name of specific supplemental file from gse to downloads}
}
\description{
Takes a GSE and a corresponding GEO supplementary file name
and downloads that file to a specified directory
}
