% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract-characteristic.R
\name{geograbr.extract.characteristic}
\alias{geograbr.extract.characteristic}
\title{Extract a characteristic field in GEO sample data}
\usage{
geograbr.extract.characteristic(x, pattern)
}
\arguments{
\item{x}{GEO characteristc field, say as returned by 
\code{geograbr.get.samples}}

\item{pattern}{for grep search}
}
\description{
Pattern searches (greps) for a particular pattern in a data object with 
sample
}
