% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{geograbr.get.data}
\alias{geograbr.get.data}
\title{Get GEO data}
\usage{
geograbr.get.data(filename, path = NULL, geo.sites = NULL,
  FUN = geo.read.gse.matrix)
}
\arguments{
\item{filename}{}
}
\description{
Grabs the expression/microarray/etc. data contained in the GEO series
 matrix files
}
