% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage-db.R
\name{geograbr.query.db.gses}
\alias{geograbr.query.db.gses}
\title{Query the GEOmetadb sql directory for GSEs}
\usage{
geograbr.query.db.gses(sql, db.path = ".")
}
\arguments{
\item{sql}{an sql expression to query the GEOmetadb database}

\item{db.path}{path where GEOmetadb.sqlite is stored}
}
\description{
Uses an sql expression to query the GEOmetadb database and returns all GSE numbers meeting the search criteria
}
