% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{geograbr.read.gse.matrix}
\alias{geograbr.read.gse.matrix}
\title{Read GEO gse series matrix}
\usage{
geograbr.read.gse.matrix(filename, data = TRUE)
}
\description{
Interface for reading a gse series matrix file as downloaded by 
\code{geograbr.download.series.files}
}
