% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{read.until}
\alias{read.until}
\title{Read lines from a connection until a pattern is matched}
\usage{
read.until(pattern, con, n = 100, ...)
}
\arguments{
\item{pattern}{Character pattern to search for using \code{\link{grep}()}.}

\item{con}{A connection object or filename.}

\item{n}{The starting number of lines to read.  Each time the pattern
is not identified, the number of lines is doubled.}

\item{...}{Additional arguments to \code{\link{readLines}()}.}
}
\value{
The initial lines read from the connection with at least
one matching the pattern or all lines available from the connection
if the pattern was not matched.
}
\description{
Read lines from a connection until a pattern is matched
}
